Drop 01_desktop-path.patch
The pupose of this patch was to force mpv.desktop to call /usr/bin/mpv
instead of mpv from $PATH.

I do not think this is a good idea since it makes 'overriding' the
behavior of mpv harder for a particular user. As another data point,
almost all the desktop files on my system do not use absolute paths.
This means that if a user does mess their $PATH up, they are likely to
break far more stuff other mpv.